+++ src/hooks/fcm_notification/useNotifications.js
@@ -1,11 +1,19 @@
+import Swal from 'sweetalert2'
+
 import { useCallback } from "react";
+import { useNavigate } from "react-router-dom";
+
 import { requestForToken, VAPID_KEY } from "@/firebase-init";
 import { getMessaging, deleteToken, getToken } from "firebase/messaging";
+
+import useAuthStore from '@/store/authStore'
 import useNotificationStore from "@/store/notificationStore";
 
 export const useNotifications = () => {
+	const navigate = useNavigate();
 	const messaging = getMessaging();
 
+	const { loginStatus } = useAuthStore.getState();
 	const { setNotification, pushEnabled, setPushEnabled } = useNotificationStore();
 
 	// 앱 로드 시 실제 권한 및 토큰 존재 여부를 확인하여 상태 동기화
@@ -29,7 +37,40 @@ export const useNotifications = () => {
 		}
 	}, [setPushEnabled]);
 
-	
+
+	const requestToPermitPushNotification = async () => {
+		// console.log("loginStatus: ", loginStatus);
+		// if (!loginStatus) return;
+		if (pushEnabled) return;
+		
+		// 아직 브라우저 알림 권한 요청한 적 없을 때 허용 요청하기
+		if (Notification.permission === 'default') {
+			await Notification.requestPermission()
+				.then((permission) => {
+					console.log("1", permission); // "granted" | "denied" | "default"
+				});
+		}
+
+		// 브라우저 알림 권한이 허용일 때 푸시 알림 허용을 위해 앱 세팅 페이지로 보내는 것 추천
+		console.log("2: ", Notification.permission)
+		if (Notification.permission === 'granted') {
+			Swal.fire({
+				title: '알림 설정',
+				text: '플래닛의 소식을 받아보시겠어요? 알림을 설정하면 에코딜, 래플 등 다양한 정보를 빠르게 얻을 수 있어요.',
+				icon: 'info',
+				showCancelButton: true,
+				confirmButtonText: '설정하기',
+				cancelButtonText: '다음에 할게요',
+			}).then((result) => {
+				if (result.isConfirmed) {
+					navigate('/my-page/settings');
+				}
+			});
+		}
+	}
+
+
+
 
 	// 권한 요청 및 토큰 발급
 	const requestPermission = async () => {
@@ -59,5 +100,5 @@ export const useNotifications = () => {
 		}
 	};
 
-	return { pushEnabled, requestPermission, revokePushToken, syncPushEnabledState };
+	return { pushEnabled, requestPermission, requestToPermitPushNotification, revokePushToken, syncPushEnabledState };
 };
